refactor: Update image upload functionality and styles in Card, ProfileConfig, Pesquisa, and Opcao components
